+++ packages/filterpane/src/epic/reducer.ts
@@ -1,12 +1,4 @@
-import {
-  Subject,
-  Subscription,
-  BehaviorSubject,
-  PartialObserver,
-  Unsubscribable,
-  Observable,
-  Subscribable,
-} from 'rxjs';
+import { Subject, Subscription, BehaviorSubject, PartialObserver, Observable } from 'rxjs';
 import { map, distinctUntilChanged, pairwise } from 'rxjs/operators';
 
 import { Action } from './epic';