+++ packages/filterpane/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.11](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.1.10...@equinor/fusion-react-filter@1.1.11) (2021-10-18)
+
+
+### Bug Fixes
+
+* **filter:** epic unused imports ([946f5dc](https://github.com/equinor/fusion-react-components/commit/946f5dc3c0963d5a1094795b17ae591cd200edd7))
+
+
+
+
+
 ## [1.1.10](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.1.9...@equinor/fusion-react-filter@1.1.10) (2021-10-13)